+++ .rubocop_todo.yml
@@ -11,11 +11,6 @@
 Metrics/AbcSize:
   Max: 84
 
-# Offense count: 2
-# Configuration parameters: CountComments, CountAsOne.
-Metrics/ClassLength:
-  Max: 189
-
 # Offense count: 9
 # Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
